run_nmtran: add compatibility handling for --maxlim=100
The default of bbi's --maxlim option used to be 100.  In that case,
bbi does not specify a -maxlim option to nmfe, leading to nmfe calling
NMTRAN with "0" as the maxlim positional argument.

bbi changed the default to 2 in b30413e8 (nonmem: pass -maxlim=2 to
nmfe by default, 2022-08-01), part of the 3.2.0 release.  bbi's
behavior remained the same when --maxlim is set to either 0 or 100
(via bbi.yaml or a command-line argument): it doesn't specify a
-maxlim option to nmfe, leading to NMTRAN getting called with "0".

Notably, if a bbi.yaml was generated with an older bbi, it by will
have nmfe_options.maxlim set to 100 by default, and bbi 3.2.0 and on
will continue to use that value.

This has two consequences for run_nmtran:

 * run_nmtran may pass an invalid argument of "100" to NMTRAN if the
   caller has that value set in the bbi.yaml

 * the run_nmtran tests fail if executed with a bbi before 3.2.0
   because 'bbi init' generates a bbi.yaml with an nmfe_options.maxlim
   value of 100 rather than 2.

Teach parse_nmtran_args to map 100 to 0, and update the run_nmtran
test to expect a default of 100 if the bbi version is before 3.2.0.