add_dofv: don't error on non-existent models
add_dofv calls read_model on the run log model paths and the based_on
model paths.  These calls will trigger an error if a model doesn't
exist.

The most likely scenario for a non-existent model is the model's first
based_on value referencing a model that doesn't exist.  It's also
possible for a row in the run or summary log to reference a model that
was deleted _after_ the log was created but before it is passed to
add_dofv.

Allow for non-existent models by replacing the read_model calls with
calls to a wrapper that demotes any error to a warning.  (Add a new
wrapper rather than using safe_read_model because safe_read_model
catches only the "missing required keys" error thrown by
create_run_log_object.)