find_models: avoid confusing "All models excluded by filter" warning
When find_models doesn't find any models, it emits a "Found no valid
model YAML files in ..." warning and then the downstream code triggers
the "All models excluded by filter" warning.

The second warning isn't appropriate in this case because the model
list was empty before filtering on .include.

None of the downstream code is relevant for the "no models" case.
Return early to avoid the invalid warning.